requirements.txt

psutil
requests
PyYAML
prometheus_client
